+++ src/plot_recipes/UncertainDataPlotRecipes.jl
@@ -4,6 +4,7 @@ using Reexport
     import ..Resampling.resample
     using RecipesBase
 
+    include("recipes_certainvalues.jl")
     include("recipes_uncertainvalues_theoretical.jl")
     include("recipes_uncertainvalues_kde.jl")
     include("recipes_uncertaindatasets.jl")