+++ src/plot_recipes/recipes_uncertain_index_and_value.jl
@@ -32,11 +32,7 @@ end
     quants_values::Vector{Float64} = [0.33, 0.67])
 
     
-    legend --> false
     n_vals = length(uvals)
-
-    xlabel --> "Index"
-    ylabel --> "Value"
 
     for i = 1:n_vals
         idx, val = uvals[i]
@@ -52,6 +48,7 @@ end
 
         @series begin 
             #seriescolor --> :black
+            label --> ""
             xerr --> ([med_idx - lower_idx], [upper_idx - med_idx])
             yerr --> ([med_val - lower_val], [upper_val - med_val])
             [med_idx], [med_val]