+++ src/plot_recipes/recipes_uncertaindatasets.jl
@@ -7,9 +7,6 @@ import ..UncertainStatistics: mean, std, median, quantile
 @recipe function f(udata::AbstractUncertainValueDataset, quants::Vector{Float64} = [0.33, 0.67])
     n_points = length(udata)
     
-    legend --> false
-    xlabel --> "Index"
-    ylabel --> "Value"
     
     for i = 1:n_points
         med = median(udata[i], 10000)
@@ -18,8 +15,10 @@ import ..UncertainStatistics: mean, std, median, quantile
 
         @series begin 
             seriescolor --> :black
+            label --> ""
             yerr --> ([med - lower], [upper - med])
             [i], [med]
+
         end
     end