+++ src/resampling/resample_uncertaindataset_withconstraint.jl
@@ -1,4 +1,5 @@
-import ..UncertainDatasets.UncertainDataset
+import ..UncertainDatasets: 
+	AbstractUncertainValueDataset
 import ..SamplingConstraints:
 	NoConstraint,
 	TruncateLowerQuantile,
@@ -9,112 +10,128 @@ import ..SamplingConstraints:
 # Draw realisations of the uncertain dataset according to the distributions
 # of the uncertain values comprising it.
 ##########################################################################
-
 """
-	resample(uv::UncertainDataset, constraint::NoConstraint) -> Vector{Float64}
+	resample(uv::DT, constraint::SCT) -> Vector{Float64}
 
 Draw a realisation of an `UncertainDataset` where each uncertain value
 is truncated according to `constraint`. In the case of `NoConstraint`,
 no trucation is performed and the whole distribution is sampled.
 """
-function resample(uv::UncertainDataset, constraint::NoConstraint)
+function resample(uv::DT, 
+		constraint::SCT) where {DT <: AbstractUncertainValueDataset,
+								SCT <: SamplingConstraint}
 	L = length(uv)
 	[resample(uv.values[i], constraint) for i in 1:L]
 end
 
 """
-	resample(uv::UncertainDataset, n::NoConstraint,
+	resample(uv::DT, constraint::NoConstraint) -> Vector{Float64}
+
+Draw a realisation of an `UncertainDataset` where each uncertain value
+is truncated according to `constraint`. In the case of `NoConstraint`,
+no trucation is performed and the whole distribution is sampled.
+"""
+function resample(uv::DT, 
+		constraint::NoConstraint) where {DT <: AbstractUncertainValueDataset}
+	L = length(uv)
+	[resample(uv.values[i], constraint) for i in 1:L]
+end
+
+"""
+	resample(uv::DT, n::NoConstraint,
 		n::Int) -> Vector{Vector{Float64}}
 
 Draw `n` realisation of an `UncertainDataset` where each uncertain value
 is truncated according to `constraint`. In the case of `NoConstraint`,
 no trucation is performed and the whole distribution is sampled.
 """
-function resample(uv::UncertainDataset, constraint::NoConstraint, n::Int)
+function resample(uv::DT, constraint::NoConstraint, 
+		n::Int) where {DT <: AbstractUncertainValueDataset}
 	L = length(uv)
 	[[resample(uv.values[i], constraint) for i in 1:L] for k = 1:n]
 end
 
 """
-	resample(uv::UncertainDataset,
+	resample(uv::DT,
 		constraint::TruncateLowerQuantile) -> Vector{Float64}
 
 Draw a realisation of an `UncertainDataset` where each uncertain value
 is truncated according to `constraint`. In the case of `TruncateLowerQuantile`,
 the supports of the distributions are truncated below at some quantile.
 """
-function resample(uv::UncertainDataset,
-		constraint::TruncateLowerQuantile)
+function resample(uv::DT,
+		constraint::TruncateLowerQuantile) where {DT <: AbstractUncertainValueDataset}
 	L = length(uv)
 	[resample(uv.values[i], constraint) for i in 1:L]
 end
 
 """
-	resample(uv::UncertainDataset, constraint::TruncateLowerQuantile,
+	resample(uv::DT, constraint::TruncateLowerQuantile,
 		n::Int) -> Vector{Vector{Float64}}
 
 Draw `n` realisation of an `UncertainDataset` where each uncertain value
 is truncated according to `constraint`. In the case of `TruncateLowerQuantile`,
 the supports of the distributions are truncated below at some quantile.
 """
-function resample(uv::UncertainDataset, constraint::TruncateLowerQuantile,
-	n::Int)
+function resample(uv::DT, constraint::TruncateLowerQuantile,
+		n::Int) where {DT <: AbstractUncertainValueDataset}
 	L = length(uv)
 	[[resample(uv.values[i], constraint) for i in 1:L] for k = 1:n]
 end
 
 """
-	resample(uv::UncertainDataset,
+	resample(uv::DT,
 		constraint::TruncateUpperQuantile) -> Vector{Float64}
 
 Draw a realisation of an `UncertainDataset` where each uncertain value
 is truncated according to `constraint`. In the case of `TruncateLowerQuantile`,
 the supports of the distributions are truncated above at some quantile.
 """
-function resample(uv::UncertainDataset,
-		constraint::TruncateUpperQuantile)
+function resample(uv::DT,
+		constraint::TruncateUpperQuantile) where {DT <: AbstractUncertainValueDataset}
 	L = length(uv)
 	[resample(uv.values[i], constraint) for i in 1:L]
 end
 
 """
-	resample(uv::UncertainDataset, constraint::TruncateUpperQuantile,
+	resample(uv::DT, constraint::TruncateUpperQuantile,
 		n::Int) -> Vector{Vector{Float64}}
 
 Draw `n` realisation of an `UncertainDataset` where each uncertain value
 is truncated according to `constraint`. In the case of `TruncateUpperQuantile`,
 the supports of the distributions are truncated above at some quantile.
 """
-function resample(uv::UncertainDataset, constraint::TruncateUpperQuantile,
-		n::Int)
+function resample(uv::DT, constraint::TruncateUpperQuantile,
+		n::Int) where {DT <: AbstractUncertainValueDataset}
 	L = length(uv)
 	[[resample(uv.values[i], constraint) for i in 1:L] for k = 1:n]
 end
 
 
 """
-	resample(uv::UncertainDataset,
+	resample(uv::DT,
 		constraint::TruncateQuantiles) -> Vector{Float64}
 
 Draw a realisation of an `UncertainDataset` where each uncertain value
 is truncated according to `constraint`. In the case of `TruncateLowerQuantile`,
 the supports of the distributions are truncated at some quantile range.
 """
-function resample(uv::UncertainDataset,
-		constraint::TruncateQuantiles)
+function resample(uv::DT,
+		constraint::TruncateQuantiles) where {DT <: AbstractUncertainValueDataset}
 	L = length(uv)
 	[resample(uv.values[i], constraint) for i in 1:L]
 end
 
 """
-	resample(uv::UncertainDataset, n::Int,
+	resample(uv::DT, n::Int,
 		constraint::TruncateUpperQuantile) -> Vector{Vector{Float64}}
 
 Draw `n` realisation of an `UncertainDataset` where each uncertain value
 is truncated according to `constraint`. In the case of `TruncateUpperQuantile`,
 the supports of the distributions are truncated above at some quantile.
 """
-function resample(uv::UncertainDataset, constraint::TruncateQuantiles, n::Int)
+function resample(uv::DT, constraint::TruncateQuantiles, 
+		n::Int) where {DT <: AbstractUncertainValueDataset}
 	L = length(uv)
 	[[resample(uv.values[i], constraint) for i in 1:L] for k = 1:n]
 end