+++ src/resampling/resampling_vector_uncertainvalues.jl
@@ -1,0 +1,8 @@
+
+function resample(uvals::Vector{AbstractUncertainValue}) 
+    [resample(uvals[i]) for i = 1:n]
+end
+
+function resample(uvals::Vector{AbstractUncertainValue}, n::Int) 
+    [resample(uvals[i], n) for i = 1:n]
+end