+++ src/sampling_constraints/ordered_sequences/ordered_sequence_algorithms.jl
@@ -11,7 +11,7 @@ abstract type OrderedSamplingAlgorithm end
 An ordered sampling algorithm indicating that values should be 
 treated consecutively from start to finish of the dataset.
 """
-struct OrderedSamplingStartToEnd <: OrderedSamplingAlgorithm end 
+struct StartToEnd <: OrderedSamplingAlgorithm end 
 
 """
     OrderedSamplingEndToStart
@@ -19,7 +19,7 @@ struct OrderedSamplingStartToEnd <: OrderedSamplingAlgorithm end
 An ordered sampling algorithm indicating that the values should be 
 treated consecutively from the end to the start of the dataset.
 """
-struct OrderedSamplingEndToStart <: OrderedSamplingAlgorithm end
+struct EndToStart <: OrderedSamplingAlgorithm end
 
 """
     OrderedSamplingMidpointOutwards
@@ -28,7 +28,7 @@ An ordered sampling algorithm indicating that the values should be
 divided into two groups, separating the values at some midpoint 
 of the dataset. The two groups of values are then treated separately.
 """
-struct OrderedSamplingMidpointOutwards <: OrderedSamplingAlgorithm
+struct MidpointOutwards <: OrderedSamplingAlgorithm
     midpoint_idx::Int
 end
 
@@ -40,7 +40,7 @@ divided into multiple (`n_chunks`) groups. The groups of values
 are then treated separately, treating values from the start to the end of 
 each group.
 """
-struct OrderedSamplingChunksForwards <: OrderedSamplingAlgorithm
+struct ChunksForwards <: OrderedSamplingAlgorithm
     n_chunks::Int
 end
 
@@ -52,14 +52,14 @@ divided into multiple (`n_chunks`) groups. The groups of values
 are then treated separately, treating values from the end to the start of 
 each group.
 """
-struct OrderedSamplingChunksBackwards <: OrderedSamplingAlgorithm
+struct ChunksBackwards <: OrderedSamplingAlgorithm
     n_chunks::Int
 end
 
 export 
 OrderedSamplingAlgorithm, 
-OrderedSamplingStartToEnd,
-OrderedSamplingEndToStart,
-OrderedSamplingMidpointOutwards,
-OrderedSamplingChunksForwards,
-OrderedSamplingChunksBackwards
+StartToEnd,
+EndToStart,
+MidpointOutwards,
+ChunksForwards,
+ChunksBackwards