+++ test/uncertain_values/test_certain_values.jl
@@ -1,0 +1,6 @@
+
+x = 3
+y = 3.3 
+
+@test UncertainValue(x) isa CertainValue
+@test UncertainValue(y) isa CertainValue